[SPARK-44283][CONNECT] Move Origin to SQL/API
### What changes were proposed in this pull request?
This PR move Origin and SQLQueryContext to sql/api.

### Why are the changes needed?
This change allows us to move ParseException to sql/api.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

Closes #41837 from hvanhovell/SPARK-44283.

Authored-by: Herman van Hovell <herman@databricks.com>
