+++ sql/api/src/main/scala/org/apache/spark/sql/catalyst/trees/SQLQueryContext.scala
